sf1xx: move to PX4Rangefinder and cleanup
