+++ src/drivers/distance_sensor/sf0x/CMakeLists.txt
@@ -1,6 +1,6 @@
 ############################################################################
 #
-#   Copyright (c) 2015 PX4 Development Team. All rights reserved.
+#   Copyright (c) 2015-2020 PX4 Development Team. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
@@ -31,7 +31,7 @@
 #
 ############################################################################
 px4_add_module(
-	MODULE drivers__sf0x
+	MODULE drivers__distance_sensor__sf0x
 	MAIN sf0x
 	COMPILE_FLAGS
 		-Wno-cast-align # TODO: fix and enable
