+++ src/drivers/distance_sensor/sf1xx/CMakeLists.txt
@@ -1,6 +1,6 @@
 ############################################################################
 #
-#   Copyright (c) 2015 PX4 Development Team. All rights reserved.
+#   Copyright (c) 2015-2020 PX4 Development Team. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
@@ -31,12 +31,13 @@
 #
 ############################################################################
 px4_add_module(
-	MODULE drivers__sf1xx
+	MODULE drivers__distance_sensor__sf1xx
 	MAIN sf1xx
-	COMPILE_FLAGS
-		-Wno-cast-align # TODO: fix and enable
 	SRCS
-		sf1xx.cpp
+		SF1XX.cpp
+		SF1XX.hpp
+		sf1xx_main.cpp
 	DEPENDS
+		drivers_rangefinder
+		px4_work_queue
 	)
-
