+++ src/drivers/distance_sensor/sf1xx/SF1XX.hpp
@@ -1,6 +1,6 @@
 /****************************************************************************
  *
- *   Copyright (c) 2014-2019 PX4 Development Team. All rights reserved.
+ *   Copyright (c) 2013-2020 PX4 Development Team. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -32,11 +32,13 @@
  ****************************************************************************/
 
 /**
- * @file SF0X.hpp
- * @author Lorenz Meier <lm@inf.ethz.ch>
- * @author Greg Hulands
+ * @file SF1XX.hpp
  *
- * Driver for the Lightware SF0x laser rangefinder series
+ * @author ecmnet <ecm@gmx.de>
+ * @author Vasily Evseenko <svpcom@gmail.com>
+ *
+ * Driver for the Lightware SF1xx lidar range finder series.
+ * Default I2C address 0x66 is used.
  */
 
 #pragma once
@@ -44,45 +46,36 @@
 #include <px4_platform_common/px4_config.h>
 #include <px4_platform_common/px4_work_queue/ScheduledWorkItem.hpp>
 #include <lib/drivers/rangefinder/PX4Rangefinder.hpp>
+#include <drivers/device/i2c.h>
 #include <drivers/drv_hrt.h>
 #include <lib/perf/perf_counter.h>
 
-#include "sf0x_parser.h"
+/* Configuration Constants */
+#define SF1XX_BASEADDR		0x66
 
-class SF0X : public px4::ScheduledWorkItem
+class SF1XX : public device::I2C, public px4::ScheduledWorkItem
 {
 public:
-	SF0X(const char *port, uint8_t rotation = distance_sensor_s::ROTATION_DOWNWARD_FACING);
-	~SF0X() override;
+	SF1XX(int bus, int address = SF1XX_BASEADDR, uint8_t rotation = distance_sensor_s::ROTATION_DOWNWARD_FACING);
+	~SF1XX() override;
 
-	int 			init();
-	void				print_info();
+	int init() override;
+	void print_info();
 
 private:
+	void start();
+	void stop();
+	void Run() override;
+	int measure();
+	int collect();
 
-	void				start();
-	void				stop();
-	void				Run() override;
-	int				measure();
-	int				collect();
+	PX4Rangefinder _px4_rangefinder;
 
+	int _interval{100000};
+	bool _collect_phase{false};
+	unsigned _consecutive_fail_count{0};
+	hrt_abstime _last_read{0};
 
-	PX4Rangefinder                  _px4_rangefinder;
-
-	char 				_port[20] {};
-	int         		        _interval{100000};
-	bool				_collect_phase{false};
-	int				_fd{-1};
-	char				_linebuf[10] {};
-	unsigned			_linebuf_index{0};
-	enum SF0X_PARSE_STATE		_parse_state {SF0X_PARSE_STATE0_UNSYNC};
-	hrt_abstime			_last_read{0};
-
-	unsigned			_consecutive_fail_count{0};
-
-	perf_counter_t			_sample_perf;
-	perf_counter_t			_comms_errors;
-
-
-
+	perf_counter_t _sample_perf;
+	perf_counter_t _comms_errors;
 };
